changed instance_label field in all structs to uint32

+++ global_segment_map/include/global_segment_map/common.h
@@ -15,7 +15,7 @@ const uint16_t BackgroundLabel = 0u;
 // Voxblox++ custom types.
 typedef uint16_t Label;
 typedef uint16_t LabelConfidence;
-typedef uint16_t InstanceLabel;
+typedef uint32_t InstanceLabel;
 typedef uint8_t SemanticLabel;
 
 typedef std::vector<Label> Labels;
@@ -42,7 +42,7 @@ struct PointSurfelLabel {
   PCL_ADD_POINT4D;
   PCL_ADD_NORMAL4D;
   PCL_ADD_RGB;
-  uint32_t label;
+  uint16_t label;
 
   EIGEN_MAKE_ALIGNED_OPERATOR_NEW
 } EIGEN_ALIGN16;
@@ -51,7 +51,7 @@ struct PointSurfelSemanticInstance {
   PCL_ADD_POINT4D;
   PCL_ADD_NORMAL4D;
   PCL_ADD_RGB;
-  uint8_t instance_label;
+  uint32_t instance_label;
   SemanticLabel semantic_label;
 
   EIGEN_MAKE_ALIGNED_OPERATOR_NEW
@@ -87,7 +87,7 @@ POINT_CLOUD_REGISTER_POINT_STRUCT(
 POINT_CLOUD_REGISTER_POINT_STRUCT(
     voxblox::PointSemanticInstanceType,
     (float, x, x)(float, y, y)(float, z, z)(float, rgb, rgb)(
-        std::uint8_t, instance_label,
+        std::uint32_t, instance_label,
         instance_label)(voxblox::SemanticLabel, semantic_label, semantic_label))
 
 POINT_CLOUD_REGISTER_POINT_STRUCT(
